refactor: Use json files, removes the need for the yaml loader
And will simplify using a transpiled version of the bar
