+++ package.json
@@ -81,7 +81,6 @@
     "webpack-cli": "3.3.9",
     "webpack-merge": "4.2.2",
     "yaml-jest": "1.0.5",
-    "yaml-loader": "0.5.0"
   },
   "dependencies": {
     "cozy-client": "13.8.3",
