--- src/config/claudyActions.yaml
@@ -1,14 +1,0 @@
-desktop:
-  icon: icon-laptop.svg
-  link:
-    type: external
-
-mobile:
-  icon: icon-phone.svg
-  link:
-    type: external
-
-support:
-  icon: icon-question-mark.svg
-  link:
-    type: external
