+++ src/config/persistWhitelist.json
@@ -1,0 +1,4 @@
+[
+  "apps",
+  "context"
+]
