--- src/config/persistWhitelist.yaml
@@ -1,2 +1,0 @@
-- apps
-- context
